[add] #7 display_map.dartにslide_up_panelを追加
