+++ src/lib/view/display_signing.dart
@@ -1,0 +1,262 @@
+import 'package:flutter/material.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import '../firebase_options.dart';
+import '../reset_password_screen.dart';
+import '../validation.dart';
+
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp(
+    options: DefaultFirebaseOptions.currentPlatform,
+  );
+  runApp(const LoginSample());
+}
+
+class LoginSample extends StatelessWidget {
+  const LoginSample({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const MaterialApp(
+      title: 'Login Sample',
+      home: LoginPage(),
+    );
+  }
+}
+
+class LoginPage extends StatefulWidget {
+  const LoginPage({super.key});
+
+  @override
+  State<LoginPage> createState() => _LoginPageState();
+}
+
+class _LoginPageState extends State<LoginPage> {
+  // メッセージ表示用
+  String infoText = '';
+  // 入力したメールアドレス・パスワード
+  String email = '';
+  String password = '';
+  final ValidationCheck _validationCheck = ValidationCheck();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Container(
+          padding: const EdgeInsets.all(24),
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              TextFormField(
+                // decoration: const InputDecoration(labelText: 'メールアドレス'),
+                decoration: InputDecoration(
+                  labelText: 'メールアドレス',
+                  labelStyle: TextStyle(
+                    color: Colors.black,
+                    fontSize: 16,
+                  ),
+                  hintText: 'xxxxx.xxxxx@xxxx.xxx',
+                  hintStyle: TextStyle(
+                    color: Colors.grey[500],
+                    fontSize: 12,
+                    fontWeight: FontWeight.w400,
+                    fontStyle: FontStyle.normal,
+                  ),
+                  icon: Icon(
+                    FontAwesomeIcons.solidEnvelope,
+                    color: Colors.black,
+                  ),
+                  counterText: '${email.length}',
+                  counterStyle: TextStyle(
+                    color: Colors.black,
+                  ),
+                  enabledBorder: UnderlineInputBorder(
+                    borderSide: BorderSide(
+                      color: Colors.black,
+                    ),
+                  ),
+                  focusedBorder: UnderlineInputBorder(
+                    borderSide: BorderSide(
+                      color: Colors.black,
+                    ),
+                  ),
+                  border: UnderlineInputBorder(
+                    borderSide: BorderSide(
+                      color: Colors.black,
+                    ),
+                  ),
+                  errorStyle: TextStyle(
+                    color: Colors.red,
+                  ),
+                ),
+                autovalidateMode: AutovalidateMode.onUserInteraction,
+                validator: (value) {
+                  return _validationCheck.emailAddressCheck(value!);
+                },
+                onChanged: (String value) {
+                  setState(() {
+                    email = value;
+                  });
+                },
+              ),
+              // パスワード入力
+              TextFormField(
+                // decoration: const InputDecoration(labelText: 'パスワード'),
+                decoration: InputDecoration(
+                  labelText: 'パスワード',
+                  labelStyle: TextStyle(
+                    color: Colors.black,
+                    fontSize: 16,
+                  ),
+                  hintText: 'パスワード(英数記号6文字以上30文字以内)',
+                  hintStyle: TextStyle(
+                    color: Colors.grey[500],
+                    fontSize: 12,
+                    fontWeight: FontWeight.w400,
+                    fontStyle: FontStyle.normal,
+                  ),
+                  icon: Icon(
+                    FontAwesomeIcons.lock,
+                    color: Colors.black,
+                  ),
+                  counterText: '${password.length}',
+                  counterStyle: TextStyle(
+                    color: Colors.black,
+                  ),
+                  enabledBorder: UnderlineInputBorder(
+                    borderSide: BorderSide(
+                      color: Colors.black,
+                    ),
+                  ),
+                  focusedBorder: UnderlineInputBorder(
+                    borderSide: BorderSide(
+                      color: Colors.black,
+                    ),
+                  ),
+                  border: UnderlineInputBorder(
+                    borderSide: BorderSide(
+                      color: Colors.black,
+                    ),
+                  ),
+                  errorStyle: TextStyle(
+                    color: Colors.red,
+                  ),
+                ),
+                obscureText: true,
+                autovalidateMode: AutovalidateMode.onUserInteraction,
+                validator: (value) {
+                  return _validationCheck.passwordCheck(value!);
+                },
+                onChanged: (String value) {
+                  setState(() {
+                    password = value;
+                  });
+                },
+              ),
+              Container(
+                padding: const EdgeInsets.all(8),
+                // メッセージ表示
+                child: Text(infoText),
+              ),
+              SizedBox(
+                width: double.infinity,
+                // ユーザー登録ボタン
+                child: ElevatedButton(
+                  child: const Text('ユーザー登録'),
+                  onPressed: () async {
+                    try {
+                      // メール/パスワードでユーザー登録
+                      final FirebaseAuth auth = FirebaseAuth.instance;
+                      await auth.createUserWithEmailAndPassword(
+                        email: email,
+                        password: password,
+                      );
+                      // ユーザー登録に成功した場合
+                      setState(() {
+                        infoText = "登録に成功しました！";
+                      });
+                    } catch (e) {
+                      // ユーザー登録に失敗した場合
+                      setState(() {
+                        infoText = "登録に失敗しました：${e.toString()}";
+                      });
+                    }
+                  },
+                ),
+              ),
+              SizedBox(
+                width: double.infinity,
+                // ログインボタン
+                child: ElevatedButton(
+                  child: const Text('ログイン'),
+                  onPressed: () async {
+                    try {
+                      // メール/パスワードでログイン
+                      final FirebaseAuth auth = FirebaseAuth.instance;
+                      await auth.signInWithEmailAndPassword(
+                        email: email,
+                        password: password,
+                      );
+                      // ユーザー登録に成功した場合
+                      setState(() {
+                        infoText = "ログインに成功しました！";
+                      });
+                    } catch (e) {
+                      // ユーザー登録に失敗した場合
+                      setState(() {
+                        infoText = "ログインに失敗しました：${e.toString()}";
+                      });
+                    }
+                  },
+                ),
+              ),
+              SizedBox(
+                width: double.infinity,
+                // ログアウトボタン
+                child: ElevatedButton(
+                  child: const Text('ログアウト'),
+                  onPressed: () async {
+                    try {
+                      // ログアウト
+                      await FirebaseAuth.instance.signOut();
+                      // ユーザー登録に成功した場合
+                      setState(() {
+                        infoText = "ログアウトしました";
+                      });
+                    } catch (e) {
+                      // ユーザー登録に失敗した場合
+                      setState(() {
+                        infoText = "ログアウトに失敗しました：${e.toString()}";
+                      });
+                    }
+                  },
+                ),
+              ),
+              TextButton(
+                onPressed: () {
+                  Navigator.of(context).push(
+                    MaterialPageRoute(
+                      builder: (context) {
+                        return const ResetPasswordScreen();
+                      },
+                    ),
+                  );
+                },
+                child: const Text(
+                  'パスワードを忘れたら',
+                  style: TextStyle(
+                    color: Colors.red,
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
